arch,toolchain,xcompile: Use the compiler frontend as a linker
Code generation when using LTO is not done until after linking, so the
compiler must be invoked at all linking stages instead of the linker.
Replace $(LD) with $(CC), and prefix all linker-specific flags with -Wl
and replace the spaces by commas so they will be passed to the linker.

Pass CFLAGS_common to SMM targets, so `-m32` gets passed, results in

    clang: error: argument unused during compilation: '-nostdinc' [-Werror,-Wunused-command-line-argument]

Pass CFLAGS_smm and CFLAGS_smmstub.

Change-Id: Ia551cf84c8f6eb3b010450fc572876a318a257d1
Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Signed-off-by: Paul Menzel <pmenzel@molgen.mpg.de>